### 2020-12-07 #2891 @whedon
[REVIEW]: Learning Simulator: A simulation software for animal and human learning
**Submitting author:** @markusrobertjonsson (<a href="http://orcid.org/0000-0003-1242-3599">Markus Jonsson</a>)
**Repository:** <a href="https://github.com/learningsimulator/learningsimulator" target ="_blank">https://github.com/learningsimulator/learningsimulator</a>
**Version:** v1.0.1
