### 2020-12-07 a7a1c85
Fix of issue 125, add test, and robustifying other tests.
