### 2020-12-07 df9e40c
tolman mechanism: first version of z() estimation + documents under dev/
