+++ joss_manuscript/paper.bib
@@ -39,6 +39,7 @@
   number={11},
   pages={160734},
   year={2016},
+  doi={10.1098/rsos.160734},
   publisher={The Royal Society}
 }
 
@@ -46,6 +47,7 @@
   title={Carrots and sticks: {P}rinciples of animal training},
   author={McGreevy, Paul and Boakes, Robert},
   year={2011},
+  doi={10.1111/j.1751-0813.2010.00653.x},
   publisher={Darlington Press}
 }
 
@@ -57,6 +59,7 @@
   number={3},
   pages={229--234},
   year={1999},
+  doi={10.1016/s0899-9007(98)00192-0},
   publisher={Elsevier}
 }
 
@@ -64,6 +67,7 @@
   title={Clinical applications of learning theory},
   author={Haselgrove, Mark and Hogarth, Lee},
   year={2013},
+  doi={10.4324/9780203803509},
   publisher={Psychology Press}
 }
 
@@ -82,6 +86,7 @@
   number={6},
   pages={1291--1307},
   year={2012},
+  doi={10.1109/tsmcc.2012.2218595},
   publisher={IEEE}
 }
 
@@ -93,6 +98,7 @@
   number={3},
   pages={181777},
   year={2019},
+  doi={10.1101/446906},
   publisher={The Royal Society}
 }
 
@@ -104,6 +110,7 @@
   number={11},
   pages={180778},
   year={2018},
+  doi={10.1098/rsos.180778},
   publisher={The Royal Society}
 }
 
@@ -133,6 +140,7 @@
   number={6},
   pages={161011},
   year={2017},
+  doi={10.1098/rsos.161011},
   publisher={The Royal Society Publishing}
 }
 
@@ -144,6 +152,7 @@
   number={2},
   pages={442--449},
   year={2008},
+  doi={10.3758/brm.40.2.442},
   publisher={Springer}
 }
 
@@ -155,6 +164,7 @@
   number={1},
   pages={1--26},
   year={2010},
+  doi={10.3758/lb.38.1.1},
   publisher={Springer}
 }
 
@@ -166,6 +176,7 @@
   number={1},
   pages={346--355},
   year={2012},
+  doi={10.1016/j.cmpb.2012.02.004},
   publisher={Elsevier}
 }
 
@@ -186,6 +197,7 @@
   number={1},
   pages={29--34},
   year={2009},
+  doi={10.3758/brm.41.1.29},
   publisher={Springer}
 }
 
@@ -197,6 +209,7 @@
   number={2},
   pages={435--441},
   year={2008},
+  doi={10.3758/brm.40.2.435},
   publisher={Springer}
 }
 
@@ -257,6 +270,7 @@
   booktitle={2009 IEEE Symposium on Adaptive Dynamic Programming and Reinforcement Learning},
   pages={177--184},
   year={2009},
+  doi={10.1109/adprl.2009.4927542},
   organization={IEEE}
 }
 
@@ -268,7 +282,8 @@
   volume={34},
   number={4},
   pages={286--295},
-  year={1977}
+  year={1977},
+  doi={10.1016/s0019-9958(77)90354-0}
 }
 
 % A-learning
@@ -278,7 +293,8 @@
   journal={Psychonomic Bulletin \& Review},
   pages={1--29},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.3758/s13423-020-01749-0}
 }
 
 @misc{QLSim,
