### 2020-12-10 33b0866
first version of tolman mechanism that can learn a single response
