+++ mechanism.py
@@ -1,5 +1,5 @@
 from math import exp
-from random import seed, random
+from random import seed, random, choice
 
 import keywords as kw
 from util import dict_inv, ParseUtil
@@ -480,35 +480,39 @@ class TolmanMechanism(Mechanism):
     def subject_reset(self):
         super().subject_reset()
         self.z = dict(self.parameters.get(kw.START_Z))  # Indexed by element-behavior-element triples
+        self.seen_stimuli = list()
         
     def learn_and_respond(self, stimulus, omit=False):        
 
-        print( "seen stimuli:", self.seen_stimuli )
-        print( "stimulus:", stimulus )
-        if stimulus not in self.seen_stimuli:
-            print( "new stimulus", stimulus )
-            self.seen_stimuli.append( stimulus )
-        print( "seen stimuli:", self.seen_stimuli )
+        if self.prev_stimulus and self.response and not omit:
+            self.learn(stimulus)
 
+        # print( "-"*20, "\nsti:", stimulus )
+
+        if stimulus not in self.seen_stimuli:
+            self.seen_stimuli.append( stimulus )
+
+        # shortcuts
         u = self.parameters.get(kw.U)
         behaviors = self.parameters.get(kw.BEHAVIORS)
-        behavior_cost = self.parameters.get(kw.BEHAVIOR_COST)
 
-        # find highest valued stimulus
-        # XXX finds only one even if there are more
+        # goal = highest valued stimulus in seen_stimuli
+        # (XXX finds only one even if there are more)
         usums = list()
         for s in self.seen_stimuli:
             usums.append( sum( u[e] for e in s.keys() ) )
         goal = self.seen_stimuli[ usums.index( max(usums) ) ]
 
-        # if we're already there, choose a random behavior
         if stimulus == goal:
-            self.v = dict(self.parameters.get(kw.START_V))
-            return super().learn_and_respond( stimulus, omit )
-
-        # build transition graph between experienced stimuli.
-        # transition probabilities are transformed into "distances" to
-        # later apply Dijkstra's algorithm
+            self.response = self._get_response(stimulus)
+            self.prev_stimulus = dict(stimulus)  # dict ok?
+            return self.response
+        
+        # build transition graph between seen_stimuli. probabilities
+        # are transformed into distances for Dijkstra's algorithm:
+        # dist = -log( prob ). the graph is just a list of
+        # connections, which requires some looping later on when we
+        # need to find a specific connection
         graph = list()
         for b in behaviors:
             for s1 in self.seen_stimuli:
@@ -517,47 +521,69 @@ class TolmanMechanism(Mechanism):
                     for e1 in s1.keys():
                         for e2 in s2.keys():
                             zsum += self.z.get( (e1,b,e2), 0 )
-                    print( s1, b, s2, zsum )
-                    if zsum>0:
-                        zsum = min( zsum, 1 ) # XXX overprediction
-                        graph.append( [ s1,b,s2,math.log1p(zsum) ] )
+                    zsum = min( zsum, 1 ) # XXX overprediction
+                    if zsum>0: # possible transition: add
+                        graph.append({ 'from':s1,'to':s2,'with':b,'prob':zsum })
                         
-        print( "graph:", graph )
+        # print( "graph:", graph )
         
-        # use Dijkstra's algorithm to find the shortest path from stimulus to goal
-        unvisited = self.seen_stimuli
-        distances = dict.fromkeys( [frozenset(s) for s in self.seen_stimuli], math.inf )
-        current = stimulus
-        distances[ frozenset(current) ] = 0
-        while len(unvisited):
-            # print( "current:", current )
-            # print( "unvisited:", unvisited )
-            # print( "distances:", distances )
-            # consider neighbors of current node
[…]
