### 2020-12-10 2e3eec9
added shortest path finding and pulled z-learning part into MapLearner class
