### 2020-12-13 63a7fa2
Fix setting of output object for Tolman mechanism.
