### 2020-12-16 5ba1ec6
Revised decision-making for TolmanMechanism. Setting alpha_z in script does not work
