+++ keywords.py
@@ -64,6 +64,7 @@ KEYWORDS = (BEHAVIORS,
             START_V,
             START_VSS,
             START_W,
+            START_Z,
             ALPHA_V,
             ALPHA_VSS,
             ALPHA_W,
