+++ mechanism.py
@@ -485,6 +485,7 @@ class MapLearner(Mechanism):
     def learn(self, stimulus):
         # shortcuts:
         alpha_z = self.parameters.get(kw.ALPHA_Z)
+        print( "alpha_z:", alpha_z )
         b = self.response
 
         # loop over all stimulus elements to update their prediction
@@ -504,8 +505,8 @@ class MapLearner(Mechanism):
             # now we can update the prediction of e2 for all elements
             # in prev_stimulus:
             for e1, i1 in self.prev_stimulus.items():
-                self.z[ (e1,b,e2) ] += alpha_z * ( zcorrect - zsum ) * i1 * i2
-
+                self.z[ (e1,b,e2) ] += alpha_z[ (e1,b,e2) ] * ( zcorrect - zsum ) * i1 * i2
+        
     def has_v(self):
         return True
 
@@ -520,40 +521,43 @@ class MapLearner(Mechanism):
 
 
 class TolmanMechanism(MapLearner):
-    def __init__(self, parameters):
-        super().__init__(parameters)
-        self.seen_stimuli = list()
-        
-    def subject_reset(self):
-        super().subject_reset()
-        self.z = dict(self.parameters.get(kw.START_Z))  # Indexed by element-behavior-element triples
-        self.seen_stimuli = list()
         
     def learn_and_respond(self, stimulus, omit=False):        
 
+        #print( "stimulus is", stimulus )
+        
+        # learn if we are told to and have enough experience:
         if self.prev_stimulus and self.response and not omit:
             self.learn(stimulus)
 
-        # print( "-"*20, "\nsti:", stimulus )
-
+        # possibly add stimulus to seen_stimuli:
         if stimulus not in self.seen_stimuli:
             self.seen_stimuli.append( stimulus )
-
+            
+        # reset v values
+        self.v = self.parameters.get(kw.START_V)
+            
         # shortcuts
         u = self.parameters.get(kw.U)
         behaviors = self.parameters.get(kw.BEHAVIORS)
+        behavior_cost = self.parameters.get(kw.BEHAVIOR_COST)
 
-        # goal = highest valued stimulus in seen_stimuli
+        # set goal to highest valued stimulus in seen_stimuli
         # (XXX finds only one even if there are more)
         usums = list()
         for s in self.seen_stimuli:
             usums.append( sum( u[e] for e in s.keys() ) )
-        goal = self.seen_stimuli[ usums.index( max(usums) ) ]
+        goal = format( self.seen_stimuli[ usums.index( max(usums) ) ] )
 
-        if stimulus == goal:
+        # if we are already at goal or no goal is found, return random
+        # behavior
+        if format(stimulus) == goal or max(usums)==0:
+            #print( "stimulus:", stimulus, "max usums:", max(usums) ) 
             self.response = self._get_response(stimulus)
             self.prev_stimulus = dict(stimulus)  # dict ok?
             return self.response
+
+        #print( "setting goal to", goal )
         
         # build transition graph between seen_stimuli. probabilities
         # are transformed into distances for Dijkstra's algorithm:
@@ -570,25 +574,27 @@ class TolmanMechanism(MapLearner):
                             zsum += self.z.get( (e1,b,e2), 0 )
                     zsum = min( zsum, 1 ) # XXX overprediction
                     if zsum>0: # possible transition: add
-                        graph.append({ 'from':s1,'to':s2,'with':b,'prob':zsum })
+                        graph.append({ 'from':format(s1),'to':format(s2),'with':b,'prob':zsum })
                         
-        # print( "graph:", graph )
+        #print( "graph:", graph )
         
         # use Dijkstra's algorithm to find a shortest path from
         # stimulus to goal, if one exists
 
[…]
