### 2020-12-17 cb68cf5
Fix bug in setting alpha_z.
