### 2020-12-25 a3ac1c8
Deleted incomplete MentalSimulator
