+++ mechanism.py
@@ -6,6 +6,8 @@ from util import dict_inv, ParseUtil
 
 import math
 
+import graph
+
 seed()
 
 
@@ -520,11 +522,39 @@ class MapLearner(Mechanism):
 
 
 class TolmanMechanism(MapLearner):
-        
+
+    # build transition graph between seen_stimuli. the graph is just a
+    # list of connections, which requires some looping later on when
+    # we need to find a specific connection
+    def build_graph( self, z ):
+        g = graph.Graph()
+        u = self.parameters.get(kw.U)
+        c = self.parameters.get(kw.BEHAVIOR_COST)
+        for b in self.parameters.get(kw.BEHAVIORS):
+            for s1 in self.seen_stimuli:
+                for s2 in self.seen_stimuli:
+                    zsum = 0
+                    usum = 0
+                    for e1 in s1.keys():
+                        for e2 in s2.keys():
+                            zsum += z.get( (e1,b,e2), 0 )
+                            usum += u[ e2 ]
+                    zsum = min( zsum, 1 ) # XXX overprediction
+                    if zsum>0: # possible transition: add
+                        g.addEdge(
+                            format(s1),
+                            format(s2),
+                            b,
+                            {
+                                'probability':zsum,
+                                'value':usum,
+                                'cost':c[b]
+                            }
+                        )
+        return g
+                        
     def learn_and_respond(self, stimulus, omit=False):        
 
-        #print( "stimulus is", stimulus )
-        
         # learn if we are told to and have enough experience:
         if self.prev_stimulus and self.response and not omit:
             self.learn(stimulus)
@@ -533,133 +563,39 @@ class TolmanMechanism(MapLearner):
         if stimulus not in self.seen_stimuli:
             self.seen_stimuli.append( stimulus )
             
-        # shortcuts
-        u = self.parameters.get(kw.U)
-        behaviors = self.parameters.get(kw.BEHAVIORS)
-        behavior_cost = self.parameters.get(kw.BEHAVIOR_COST)
+        ## reset v
+        self.v = self.parameters.get(kw.START_V)
 
-        # set goal to highest valued stimulus in seen_stimuli
-        # (XXX finds only one even if there are more)
-        usums = list()
-        for s in self.seen_stimuli:
-            usums.append( sum( u[e] for e in s.keys() ) )
-        goal = format( self.seen_stimuli[ usums.index( max(usums) ) ] )
+        g = self.build_graph( self.z )
+        paths = []
+        for dest in self.seen_stimuli:
+            if dest != stimulus:
+                paths = paths + g.find_simple_paths(
+                    format(stimulus),
+                    format(dest)
+                )
 
-        # if we are already at goal or no goal is found, return random
-        # behavior
-        if format(stimulus) == goal or max(usums)==0:
-            #print( "stimulus:", stimulus, "max usums:", max(usums) ) 
-            self.response = self._get_response(stimulus)
-            self.prev_stimulus = dict(stimulus)  # dict ok?
-            return self.response
+        # calculate path value and first behavior on each path
+        path_values = []
+        responses = []
+        for p in paths:
+            value = g.path_sum_property( p, 'value' )
+            cost = g.path_sum_property( p, 'cost' )
+            prob = g.path_product_property( p, 'probability' )
+            path_values.append( (value - cost)*prob / (len(p)-1) )
+            responses.append( p[1][1] )
 
-        #print( "setting goal to", goal )
+        # print( "path values:", path_values )
+        # print( "responses:", responses )
         
-        # build transition graph between seen_stimuli. probabilities
-        # are transformed into distances for Dijkstra's algorithm:
-        # dist = -log( prob ). the graph is just a list of
-        # connections, which requires some looping later on when we
[…]
