### 2020-12-30 662fde6
Added method to calculate path value
