### 2020-12-30 8c968ce
Simplified decision making in Tolman mechanism
