+++ docs/devdoc.txt
@@ -1,107 +1,0 @@
-******************
-Install matplotlib
-******************
-
-python -m pip install -U pip
-python -m pip install -U matplotlib
-
-
-*******************
-Run tests like this
-*******************
-
-Ubuntu (med Python 3.6):
-========================
-
-Run a specific test file:
--------------------------
-markus-laptop:~/lesim/lesim2$ python -m unittest -v tests/test_beta.py
-
-Run all test files
-------------------
-markus-laptop:~/lesim/lesim2$ python -m unittest -v tests/test_*.py
-
-
-Windows (med Python 3.8):
-=========================
-
-Run a specific test file:
--------------------------
-C:\Users\markj\lesim>C:\Python38\python.exe -m unittest -v tests/test_beta.py
-
-or, with pytest:
-
-C:\Users\markj\lesim>pytest tests/test_beta.py
-
-Run all test files (that starts with "test_")
----------------------------------------------
-Make sure there is an empty file __init__.py in the test folder
-C:\Users\markj\lesim>C:\Python38\python.exe -m unittest -v
-(-v is verbose)
-
-or, with pytest:
-
-C:\Users\markj\lesim>pytest [tests]
-
-
-*************************
-Build EXE-file on Windows
-*************************
-Start C:\Pipwin\pip-Win_1.9.exe which starts a GUI in which enter:
-    Python interpreter: C:\Python38\python.exe
-    Command: venv lesim-bundle
-Click Run - starts a command prompt.
-Make sure lesim.spec contains "datas=[('Lemur-icon.gif', '.')]"
-At the started command prompt:
-   C:\Users\markj\lesim\lesim2> C:\Python38\Scripts\pyinstaller.exe -F -w --add-data Lemur-icon.gif;. lesim.py
-   (The option -w makes sure that the created exe-file will not open a console window (stdin/out for Python))
-This creates C:\Users\markj\lesim\lesim2\dist\lesim.exe
-
-copy C:\Users\markj\lesim\lesim2\dist\lesim.exe  C:\Users\markj\lesim\learningsimulator.exe
-
-*****
-Debug
-*****
-python -m pdb my_program.py
-python -m winpdb my_program.py (Graphical debugger on Windows)
-
-Winpdb no install mode
-----------------------
-python /the/path/to/winpdb.py my_program.py
-
-
-**********
-Docs build
-**********
-
-(Requires packages sphinx and recommonmark.)
-
-In the docs folder, run
-
-make html (which is short for
-sphinx-build -b html C:\...\docs C:\...\docs\_build)
-
-
-********
-Coverage
-********
-Use Coverage.py (https://coverage.readthedocs.io/en/coverage-5.3, "pip install coverage").
-
-coverage run -m pytest [tests]
-C:\Python38\Scripts\coverage.exe run -m pytest [tests]
-
-This creates the file .coverage in the directory where the tests were run.
-
-Convert coverage data (.coverage) to (readable) html:
-C:\Python38\Scripts\coverage.exe html
-This creates the folder htmlcov in the directory where the tests were run.
-
-Generate coverage badge
-=======================
-Use coverage-badge (https://pypi.org/project/coverage-badge, "pip install coverage-badge").
-
-After above test has run and .coverage has been created:
-
-C:\Users\markj\lesim\lesim2>C:\Python38\Scripts\coverage-badge.exe -o coverage.svg
-
-This creates the file coverage.svg.
