### 2021-01-10 c32fdf1
Robustify test_trace
