### 2021-01-10 969520b
Adding test to test_trace.
