### 2021-01-12 #2891 @CatarauCorina
Hi @oliviaguest , I am about to finish my review, but I have 2 questions:
1. Any errors that I have encounter should I add them under the functionality observations?
2. Is the software paper the one found on readthedocs ?https://learningsimulator.readthedocs.io/en/latest/background.html
As I can't seem to find it anywhere else.

Thank you for your help!
