### 2021-01-12 #2891 @CatarauCorina
### Functionality
✅ Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
      OK: A list of dependencies to install  found in requirements.txt

✅ Functionality: Have the functional claims of the software been confirmed?

- Yes, but whenever I press simulate and plot I am getting an additional window not shown in the documentation called “Plot and Export Progress”. This number of this windows increases with each simulation such that in the below image I have ran 3 simulations and therefore apart from my plot window is these 3 extra windows. Unsure weather this has to be addressed

![image](https://user-images.githubusercontent.com/8167831/104338873-a6659180-54ff-11eb-8304-3adb64955cec.png)

- - “ A stimulus is specified as a comma-separated list of stimulus elements, optionally with the intensity in brackets after the stimulus element name, for example e1,e2, or e1[0.3],e2,e3[0.9]” . However when modifying the default script to add intensity to the stimulus elements I get an error. Should the intensity be added in a different place ? And if so I think it's unclear in the documentation

![image](https://user-images.githubusercontent.com/8167831/104338938-bb422500-54ff-11eb-8d3a-f878bc7007a0.png)

Here is the head of the input file
n_subjects        = 10
mechanism         = sr
behaviors         = response, no_response
stimulus_elements = stimulus[0.9],background,reward[0.3]
start_v           = -1
alpha_v           = 0.1
u                 = reward:10, default:0

-  “ n_subjects = n where n is a scalar expression evaluating to a positive integer”.  When running **python lesim.py gui** and adding a value to n_subjects < 5 it fails, when running it in script mode it performs ok.

### Documentation

✅ A statement of need  Do the authors clearly state what problems the software is designed to solve and who the target audience is?

- Yes, however it does not state the intended audience, I would add who the user persona is, in order to understand the background necessary to use it properly

✅ Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

- Yes, there is one example script for stimulus-response mechanism, though considering the complexity of the language a few more would be useful , for example one for every mechanism.

✅ Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
	- Yes, but I couldn't find how to run all of them at once, and some seemed to fail as I ran them one by one
