### 2021-01-12 #2891 @markusrobertjonsson
@oliviaguest I can respond to issues and questions raised by the reviewers as they come in, without having to wait for a final assessment, right? And I can make commits to source code and docs to address them, if needed?
