### 2021-01-21 cc0f243
Addressing JOSS reviewer comments.
