### 2021-01-22 #2891 @markusrobertjonsson
@CatarauCorina Thanks for reviewing! Below are my comments to the issues you raised.
- Regarding the Plot/Export Progress: We have noticed a few window-related problems on Mac. This seems to be yet another one. The Progress dialog box just displays the progress bar and does not affect the functionality. I have now disabled it on Mac, and made an issue to look into this. To help with the debugging: Which OSX version are you using?
- Regarding the problem with intensities: Stimulus intensities should not be specified in the parameter `stimulus_elements`, but in a `@phase`-block where it specifies the intensity of the stimulus element that is presented to the subject. The parameter `stimulus_elements` just declares the names of the elements. I have clarified the documentation.
- Regarding the problem with `n_subjects`<5: As this does not seem to be a general problem, feel free to either send me the script where the error occured, and/or submit an issue.
- Regarding the intended audience: In the README-file as well as in the Summary section on https://github.com/learningsimulator, it says "It is primarily intended for computational and behavioral biologists, ethologists, and psychologists, and it can also be useful to students and teachers in these fields." Let me know if you think this should be stated elsewhere as well.
- Regarding example usage: Since changing the mechanism merely involves setting the value of the property `mechanism` (to `ga`, `sr`, `es`, `ql`, `ac`, or `rw`) while keeping the rest of the script intact, the complexity of the scripting language will not be showcased this way. Instead we refer to the script database on the project's website https://www.learningsimulator.org/script-database. I have added a reference to the database in the doc section "A Small Usage Example".
- Regarding automated tests: There are automated tests (that are run by GitHub Actions), and the build-badge "Learning Simulator" shows whether or not they are passing. In a Python environment with `pytest` or `unittest` installed, the test suite can be run with `python -m pytest` (as in `workflow.yml`) or `python -m unittest`. To run an individual test file (for example `test_beta.py` that tests the script parameter `beta`), use `python -m pytest tests/test_beta.py` or `python -m unittest tests/test_beta.py`. (As far as I can see, the JOSS docs criteria for tests only says "An automated test suite hooked up to an external service such as Travis-CI or similar", not that there has to be instructions on how to manually execute the automated test suite.)
