### 2021-01-25 #2891 @CatarauCorina
@markusrobertjonsson , thank you for your response !

1. 10.14.5 is the version I have
2. Thank you for adding that to the documentation, is clear now , I have tested and it is working fine.
3. Ok, it's not a urgent issue, as I was saying in script mode is working fine, plus if I'm not mistaken usually in this type of experiments larger number of subjects are used.
4. That's fine - I might have missed it. Sorry !
5. That's perfect as the script database has 3 types of mechanisms, so that's great. Thanks!
6.  You are correct, I didn't see there was a CI pipeline in place, I saw it now under actions and had a look at the stages of the build pipeline and you are correct they are running fine in that environment so it is surely something to do with my machine, which is not a requirement.

Thank you for clarifying my issues, I appreciate it. Good luck with your review further !
