### 2021-01-31 8e60a56
Test plotting in a non-blocking way on Mac to address issue 86.
