### 2021-02-04 6583fda
Fix of issue 49
