### 2021-02-04 b44e015
Fix of issue 49
