### 2021-02-09 f132a98
Fix of mac rendering glitch and mac gui error.
