### 2021-02-12 #2891 @themkar
Hi @oliviaguest, I have finished my review and I am appending my comments (minor points) below.
Many thanks.

**comment #1**
_Functionality: Have the functional claims of the software been confirmed?_

I have experimented with LeSIm using the GUI and the Computer Lab document and the three example scripts. I think it achieves the simplicity to which the authors refer to in the statement in need.
The software offers good functionality for producing plots.

The GUI produced this warning in the chaining web, which can be explained better.

![image](https://github.com/themkar/myimages/raw/main/lesim.png)

It would be good to have text processing functionalities under edit- copy, paste, find (or with keyboard shortcuts). Not sure if it is my system, but only undo and redo were available.

**comment #2**
_Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution._

Yes. Tried installation on Win, MacOS, Linux Mint. It might be worth including the possibility in which python is called with python3 (no alias).

**comment #3**
_Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)._
Yes. would be good if a sample script folder was incorporated in the GUI.

**comment #4**
 _Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?_
Yes. And I especially liked the availability of a tutorial.
