### 2021-02-15 #2891 @markusrobertjonsson
@themkar Thank you for reviewing! Below are my responses to your comments.

**comment #1**
The warning is a deprecated-syntax-warning. I have edited the script to use the updated syntax, so that the warning does not appear in that script anymore.

Additional text-processing functionalities are planned (see e.g. https://github.com/markusrobertjonsson/learningsimulator/issues/6), but since it is possible for the user to use his/her favourite editor and run scripts from a command line, we have not prioritized this.

**comment #2**
I have added python3 (and pip3) to the installation instructions.

**comment #3**
The addition of sample scripts (in a "Script Library") is a very good idea and is indeed planned for the future (https://github.com/markusrobertjonsson/learningsimulator/issues/26).

**comment #4**
Great!
