### 2021-02-16 031e512
Adding missing doi in JOSS manuscript.
