+++ joss_manuscript/paper.bib
@@ -222,6 +222,7 @@
   number={5},
   pages={313},
   year={1951},
+  doi={10.1037/h0054388},
   publisher={American Psychological Association}
 }
 
