### 2021-02-17 8fee5e6
Update README.md with Zenodo DOI badge.
