### 2021-02-22 #2891 @markusrobertjonsson
@oliviaguest The title on Zenodo is now changed to match the JOSS submission.
