### 2021-02-23 #2891 @kthyng
Hi @markusrobertjonsson! I'll be helping to finalize your publication. Can you also update the metadata on your Zenodo archive so that the author list matches your JOSS paper?
